# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api.ErrorResponse:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/api.ResponseCode'
        description: response code
      error:
        description: error detail
        type: string
      errors:
        description: field errors
        items:
          $ref: '#/definitions/api.FieldError'
        type: array
      message:
        description: response message
        type: string
      request_id:
        description: request id
        type: string
      timestamp:
        description: timestamp (ms)
        type: integer
    type: object
  api.FaceCompareDTO:
    properties:
      code:
        type: integer
      comparison_results:
        properties:
          confidence:
            type: number
          confidence_exist:
            items:
              type: number
            type: array
          detection_result:
            type: string
          is_face_exist:
            type: integer
          is_same_face:
            type: integer
        type: object
      filename:
        items:
          type: string
        type: array
      msg:
        type: string
    type: object
  api.FaceCompareSuccessResponse:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/api.ResponseCode'
        description: response code
      data:
        $ref: '#/definitions/api.FaceCompareDTO'
      message:
        description: response message
        type: string
      request_id:
        description: request id
        type: string
      timestamp:
        description: timestamp (ms)
        type: integer
    type: object
  api.FaceDetectDTO:
    properties:
      code:
        type: integer
      detection_results:
        properties:
          face_num:
            type: integer
          faces_detected:
            items:
              properties:
                confidence:
                  type: number
                facial_area:
                  properties:
                    h:
                      type: integer
                    left_eye:
                      items:
                        type: integer
                      type: array
                    right_eye:
                      items:
                        type: integer
                      type: array
                    w:
                      type: integer
                    x:
                      type: integer
                    "y":
                      type: integer
                  type: object
              type: object
            type: array
          is_face_exist:
            type: integer
        type: object
      filename:
        type: string
      msg:
        type: string
    type: object
  api.FaceDetectSuccessResponse:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/api.ResponseCode'
        description: response code
      data:
        $ref: '#/definitions/api.FaceDetectDTO'
      message:
        description: response message
        type: string
      request_id:
        description: request id
        type: string
      timestamp:
        description: timestamp (ms)
        type: integer
    type: object
  api.FaceSearchDTO:
    properties:
      code:
        type: integer
      filename:
        type: string
      msg:
        type: string
      searching_results:
        properties:
          has_similar_picture:
            type: integer
          searched_similar_pictures:
            items:
              $ref: '#/definitions/api.FaceSearchItem'
            type: array
        type: object
    type: object
  api.FaceSearchItem:
    properties:
      confidence:
        type: number
      id:
        type: string
    type: object
  api.FaceSearchSuccessResponse:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/api.ResponseCode'
        description: response code
      data:
        $ref: '#/definitions/api.FaceSearchDTO'
      message:
        description: response message
        type: string
      request_id:
        description: request id
        type: string
      timestamp:
        description: timestamp (ms)
        type: integer
    type: object
  api.FieldError:
    properties:
      field:
        description: field name
        type: string
      message:
        description: error message
        type: string
      value:
        description: field value
        type: string
    type: object
...
  api.OCRPublicResult:
    properties:
      code:
        type: integer
      filename:
        type: string
      full_text:
        type: string
      msg:
        type: string
      parsing_results:
        additionalProperties:
          $ref: '#/definitions/api.OCRTextItem'
        type: object
    type: object
  api.OCRSuccessResponse:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/api.ResponseCode'
        description: response code
      data:
        $ref: '#/definitions/api.OCRPublicResult'
      message:
        description: response message
        type: string
      request_id:
        description: request id
        type: string
      timestamp:
        description: timestamp (ms)
        type: integer
    type: object
  api.OCRTextItem:
    properties:
      confidence:
        type: number
      text:
        type: string
    type: object
  api.PasswordResetRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  api.ResetPasswordConfirmRequest:
    properties:
      new_password:
        minLength: 6
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
  api.ResponseCode:
    enum:
    - 200
    - 1000
    - 1001
    - 1002
    - 1003
    - 1004
    - 1005
    - 1006
    - 1007
    - 2000
    - 2001
    - 2002
    - 2003
    - 2004
    - 2005
    - 2006
    - 2007
    - 40201
    - 5000
    - 5001
    - 5002
    - 5003
    - 5004
    type: integer
    x-enum-comments:
      CodeBadRequest: invalid request parameters
      CodeBusinessError: generic business error
      CodeConflict: resource conflict
      CodeDatabaseError: database error
      CodeEncryptionError: encryption error
      CodeFaceNotMatch: face not matched
      CodeFaceVerifyFailed: face verification failed
      CodeForbidden: forbidden
      CodeIDCardNotMatch: id card information not matched
      CodeInternalError: internal server error
      CodeInvalidParameter: invalid parameter
      CodeKYCFailed: kyc verification failed
      CodeLivenessFailed: liveness detection failed
      CodeMethodNotAllowed: method not allowed
      CodeMissingParameter: missing parameter
      CodeNotFound: resource not found
      CodeOCRFailed: ocr recognition failed
      CodePaymentRequired: quota exceeded, please upgrade your plan
      CodeServiceUnavailable: service unavailable
      CodeThirdPartyError: third party service error
      CodeTooManyRequests: too many requests
      CodeUnauthorized: unauthorized
    x-enum-descriptions:
    - ""
    - invalid request parameters
    - unauthorized
    - forbidden
    - resource not found
    - method not allowed
    - too many requests
    - invalid parameter
    - missing parameter
    - generic business error
    - ocr recognition failed
    - face verification failed
    - liveness detection failed
    - kyc verification failed
    - id card information not matched
    - face not matched
    - resource conflict
    - quota exceeded, please upgrade your plan
    - internal server error
    - database error
    - third party service error
    - service unavailable
    - encryption error
    x-enum-varnames:
    - CodeSuccess
    - CodeBadRequest
    - CodeUnauthorized
    - CodeForbidden
    - CodeNotFound
    - CodeMethodNotAllowed
    - CodeTooManyRequests
    - CodeInvalidParameter
    - CodeMissingParameter
    - CodeBusinessError
    - CodeOCRFailed
    - CodeFaceVerifyFailed
    - CodeLivenessFailed
    - CodeKYCFailed
    - CodeIDCardNotMatch
    - CodeFaceNotMatch
    - CodeConflict
    - CodePaymentRequired
    - CodeInternalError
    - CodeDatabaseError
    - CodeThirdPartyError
    - CodeServiceUnavailable
    - CodeEncryptionError
...
  api.SuccessResponse:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/api.ResponseCode'
        description: response code
      data:
        description: response data
      message:
        description: response message
        type: string
      request_id:
        description: request id
        type: string
      timestamp:
        description: timestamp (ms)
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 企业级KYC认证服务
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: KYC Service API
  version: "1.0"
paths:
...
  /api/v1/auth/password-reset/confirm:
    post:
      consumes:
      - application/json
      description: 使用令牌重置用户密码
      parameters:
      - description: 重置密码请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ResetPasswordConfirmRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: 重置密码
      tags:
      - Auth
  /api/v1/auth/password-reset/request:
    post:
      consumes:
      - application/json
      description: 发送密码重置邮件到用户邮箱
      parameters:
      - description: 密码重置请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.PasswordResetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: 请求密码重置
      tags:
      - Auth
...
  /kyc/face/compare:
    post:
      consumes:
      - multipart/form-data
      description: Upload two images for face comparison
      parameters:
      - description: Source image
        in: formData
        name: source_image
        required: true
        type: file
      - description: Target image
        in: formData
        name: target_image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.FaceCompareSuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: Face comparison
      tags:
      - Public
  /kyc/face/detect:
    post:
      consumes:
      - multipart/form-data
      description: Upload an image for face detection
      parameters:
      - description: Face image
        in: formData
        name: picture
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.FaceDetectSuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: Face detection
      tags:
      - Public
  /kyc/face/search:
    post:
      consumes:
      - multipart/form-data
      description: Upload a face image to search similar faces
      parameters:
      - description: Face image
        in: formData
        name: picture
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.FaceSearchSuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: Face search
      tags:
      - Public
...
  /kyc/ocr:
    post:
      consumes:
      - multipart/form-data
      description: Upload ID card image for OCR recognition
      parameters:
      - description: ID card image
        in: formData
        name: picture
        required: true
        type: file
      - description: Language default:thai
        enum:
        - thai
        - vietnamese
        - indonesian
        - chinese
        - english
        - tagalog
        - malay
        example: thai
        in: formData
        name: language
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.OCRSuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: OCR recognition
      tags:
      - Public
      - OCR
  /kyc/status/{request_id}:
    get:
      description: 根据请求ID查询KYC认证状态
      parameters:
      - description: 请求ID
        in: path
        name: request_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: 查询KYC状态
      tags:
      - KYC
  /kyc/verify:
    post:
      consumes:
      - multipart/form-data
      description: Upload ID card image, face image and basic info for full KYC verification
      parameters:
      - description: ID card image
        in: formData
        name: idcard_image
        required: true
        type: file
      - description: Face image
        in: formData
        name: face_image
        required: true
        type: file
      - description: Name
        in: formData
        name: name
        required: true
        type: string
      - description: ID card number
        in: formData
        name: idcard
        required: true
        type: string
      - description: Phone number
        in: formData
        name: phone
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: Complete KYC verification
      tags:
      - KYC
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"